drpcclient: add additional dial options to dial a drpc conn
`drpcclient.ClientConn` is a client connection with interceptor support. It
takes a connection so that the interceptor chain can be applied to both direct
and pooled connections.

This change adds new dialer options `WithContextDialer` and `WithTLSConfig` to
simplify the consuming code paths.

This change also introduces the `WithContextDialer` and `WithTLSConfig` options
to simplify connection dialing logic in the application (aka client) code.