drpcconn, drpcmanager: add support to send and receive grpc metadata
This fix ensures that if the outgoing context has any grpc metadata, it
is processed and sent as drpc metadata in the Invoke (unary) and
NewStream (stream) flow. On the receiving end, the fix ensures that the
server stream that processes an incoming packet duplicates any drpc
metadata into the grpc metadata in the incoming context.
This will enable us in the short-term to send and receive drpc metadata
even while the cockroach code base continues to use the grpc metadata
package to populate and receive metadata, when DRPC is enabled.